refactor to view fie
